Add files via upload
Design and Code a class for a generic product. 

The product data should include a product code, description, price and product count. 

In addition to the getter and setter method, the class should implement toString and also a formatted output of the price. From the client app create instance of 2-3 different products and use the toString method to display their data.